controlport: changing the default port number on the Thrift interface to zero.
This allows allows the OS - specifically bind() - to select an
ephemeral / unused port number.

Next, going to investigate using getsockname() to determine the port
number selected by the OS and ensure that the Thrift endpoint string
reflects the correct port number.

Also going to see if we can publish the actual hostname of the machine
from either the uname() or gethostname() functions.